requirements_rag.txt
# RAG Hukuki Chatbot Requirements

# Web Framework
flask==3.0.0
flask-cors==4.0.0
python-dotenv  # For .env file support

# PDF Processing
pypdf2==3.0.1

# NLP & Embeddings
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.0
FlagEmbedding  # For BGE-M3 model

# LLM Support
accelerate==0.24.1

# Vector Database
faiss-cpu==1.7.4

# Utilities
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
tqdm==4.66.1
rank-bm25==0.2.2  # For hybrid search
pyyaml==6.0.1  # For configuration

